fix: Import the library correctly on native and web
